+++ Includes/Device/__Sys.Device.Struct.h
@@ -18,8 +18,8 @@ typedef struct
 
 	int (*Open)   (void *Device_Args,int Mode);
 	int (*Close)  (void *Device_Args);
-	int (*Read)   (void *Device_Args,long OffSet_Pos, void *Buffer, unsigned long Size,int TimeOut);
-	int (*Write)  (void *Device_Args,long OffSet_Pos, const void *Buffer, unsigned long Size,int TimeOut);
+	int (*Read)   (void *Device_Args,long OffSet_Pos, void *Buffer, unsigned long Size,long TimeOut);
+	int (*Write)  (void *Device_Args,long OffSet_Pos, const void *Buffer, unsigned long Size,long TimeOut);
 	int (*Control)(void *Device_Args,int Cmd, unsigned long Args);
 
 }__Sys_Device_OPS_Type;
